requirements.txt
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
numpy==2.0.1
pandas==2.2.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
django-cors-headers==4.4.0
